refactor: move anchor-related to anchor_info; move checker logic to checker; add run.sh to test
