+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "easy-analysis"
+name = "solana-program-analyzer"
 version = "0.1.0"
 edition = "2024"
